:recycle: [#1792] Refactor parsing errors
